+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v0.5.0 (in development)
 -----------------------
+- `Properties` instances can now compare equal to `dict`s
 
 v0.4.0 (2017-04-22)
 -------------------