+++ CHANGELOG.md
@@ -1,11 +1,12 @@
-v0.5.0 (in development)
------------------------
+v0.5.0 (2018-09-18)
+-------------------
 - **Breaking**: Invalid `\uXXXX` escape sequences now cause an
   `InvalidUEscapeError` to be raised
 - `Properties` instances can now compare equal to `dict`s and other mapping
   types
 - Gave `Properties` a `copy` method
 - Drop support for Python 2.6 and 3.3
+- Fixed a `DeprecationWarning` in Python 3.7
 
 v0.4.0 (2017-04-22)
 -------------------